Add support for "#" meaning temporary table as in SQL server
Explanation: There was basically no other way to do it in a cross-database manner. It is way easier to give "#" a meaning and always filter it out than passing a parameter around as "#" isn't allowed in other databases. Also, "#" is quite helpful in keeping track of what is persistent and temporary.